Linking to WickedEngine
Required some compromise, as Lobster and WickedEngine headers can
not co-exist in the same compilation unit, because they both push
float2, int2, etc definitions out to the global namespace.  So no
Wicked or Lobster headers allowed in the precompiled header file.

Also had to change the WickedEngine project to use the mt runtime
(no dll), and set the ITERATOR_DEBUG_LEVEL to match what Lobster sets.
That will need to be revisited if I do script the setup for this.