Added support for simplest component, name component.
Feeling out what is involved.  Component references are interesting
because they are not supposed to be long lived, you grab them, do what
you need to do, and the drop them.  Anything that perturbs the ecs
components array can invalid the component pointer.  This is also
how it works for the WE lua bindings.

Fleshed out wisupport for more sanity checking, and to reduce the need
to call the wi_* bindings directly outside of the wisupport module.