Add more bindings, move to fixed update
The order I add the bindings in is haphazard as I learn in parallel
how to use the WickedEngine APIs in detail.

Moved the per-frame callback to the lobster code for the application
to Application::FixedUpdate.  Created wi::Application subclass as that's
going to orchestrate any other calls to Lobster.